requirements.txt
cellpose==3.1.1.1
gefpy==1.2.3
h5py==3.11.0
matplotlib==3.7.5
networkx==3.1
numba==0.58.1
numpy==1.24.3
onnxruntime_gpu==1.16.3
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
openpyxl==3.1.5
patchify==0.2.3
Pillow==10.4.0
Requests==2.32.3
scikit-image==0.21.0
scipy==1.10.1
seaborn==0.13.2
tensorflow==2.8.0
tf2onnx==1.16.1
tifffile==2023.7.10
tqdm==4.67.1
